V0.1/requirements.txt
customtkinter>=5.2.2
Pillow>=10.2.0
rawpy>=0.23.0